+++ src/Game.cs
@@ -12,12 +12,16 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Drawing;
 using System.Linq;
+using System.Reflection.Metadata;
 using CivOne.Advances;
 using CivOne.Buildings;
 using CivOne.Civilizations;
 using CivOne.Enums;
 using CivOne.IO;
+using CivOne.Leaders;
 using CivOne.Screens;
+using CivOne.Services;
+using CivOne.Services.GlobalWarming;
 using CivOne.Tasks;
 using CivOne.Tiles;
 using CivOne.Units;
@@ -155,9 +159,13 @@ namespace CivOne
 
 		internal IEnumerable<Player> Players => _players;
 
-		public void EndTurn()
+
+		internal IGlobalWarmingService globalWarmingService;
+
+		public IGlobalWarmingService GlobalWarmingService => globalWarmingService;
+		internal IGlobalWarmingScourgeService globalWarmingScourgeService;
 		{
-			foreach (Player player in _players.Where(x => !(x.Civilization is Barbarian)))
+			foreach (Player player in _players.Where(x => x.Civilization is not Barbarian))
 			{
 				player.HandleExtinction();
 			}
@@ -224,6 +232,18 @@ namespace CivOne
 				GameTask.Enqueue(Message.Help("--- Civilization Note ---", TextFile.Instance.GetGameText("HELP/HELP1")));
 			else if (Game.InstantAdvice && (Common.TurnToYear(Game.GameTurn) == -3200 || Common.TurnToYear(Game.GameTurn) == -2400))
 				GameTask.Enqueue(Message.Help("--- Civilization Note ---", TextFile.Instance.GetGameText("HELP/HELP2")));
+		}
+
+		protected void HandleGlobalWarming()
+		{
+			if (!globalWarmingService.IsGlobalWarmingOnNewTurn())
+			{
+				return;
+			}
+			
+			globalWarmingScourgeService.UnleashScourgeOfPollution();
+
+			GameTask.Enqueue(Message.Newspaper(null, "Global temperature", "rises! Icecaps melt.", "Severe Drought."));
 		}
 
 		// store last active player unit to check if a previous player move happened or a game was loaded.