+++ src/Screens/GamePlay.cs
@@ -390,7 +390,7 @@ namespace CivOne.Screens
 			_rightSideBar = Settings.RightSideBar;
 
 			_menuBar = new MenuBar(Palette);
-			_sideBar = new SideBar(Palette);
+			_sideBar = new SideBar(Palette, Game.GlobalWarmingService);
 			_gameMap = new GameMap();
 
 			CenterMapOnActiveHumanPlayerAsset();