+++ src/Services/GlobalWarming/IGlobalWarmingScourgeService.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+using CivOne.Enums;
+using CivOne.Tiles;
+
+namespace CivOne.Services.GlobalWarming
+{
+	public interface IGlobalWarmingScourgeService
+	{
+		void UnleashScourgeOfPollution();
+	}
+}