+++ src/Services/GlobalWarming/IGlobalWarmingStoreService.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+using CivOne.Tiles;
+
+namespace CivOne.Services.GlobalWarming
+{
+	public interface IGlobalWarmingStoreService
+	{
+		void Store(IGameData gameData);
+	}
+}